requirements.txt
# Core dependencies
Pillow>=10.0.0
typer[all]>=0.9.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking and linting
mypy>=1.9.0
pylint>=3.1.0
black>=24.2.0
isort>=5.13.0
flake8>=7.0.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
